requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.2
pydantic-settings==2.8.1
httpx==0.28.1
psutil==7.0.0
PyYAML==6.0.2
Jinja2==3.1.6
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
